+++ src/Revit.Linter.DiagnosticListPresenter/Views/DiagnosticListView.xaml.cs
@@ -1,14 +1,11 @@
-﻿using Revit.Linter.DiagnosticListPresenter.ViewModels;
-
 using Revit.Linter.ThemeManaging.Abstractions.Services;
 
 namespace Revit.Linter.DiagnosticListPresenter.Views;
 
 public sealed partial class DiagnosticListView
 {
-    public DiagnosticListView(IServiceProvider serviceProvider, IThemeService themeService)
+    public DiagnosticListView(IThemeService themeService)
     {
-        ServiceLocator.Initialize(serviceProvider);
         InitializeComponent();
         themeService.Register(this);
     }