+++ src/Revit.Linter.RunDiagnosticPresenter/DI/ServiceCollectionExtensions.cs
@@ -1,5 +1,5 @@
 using Microsoft.Extensions.DependencyInjection;
-using MVVM.DependencyInjection;
+using Revit.Linter.RunDiagnosticPresenter.ViewModels;
 using Revit.Linter.RunDiagnosticPresenter.Views;
 
 namespace Revit.Linter.RunDiagnosticPresenter.DI;
@@ -9,7 +9,12 @@ public static class ServiceCollectionExtensions
     extension(IServiceCollection services)
     {
         public IServiceCollection AddRunDiagnosticModule()
-            => services.AddView<RunDiagnosticView>(ServiceLifetime.Singleton)
+            => services
+                .AddSingleton<RunDiagnosticViewModel>()
+                .AddTransient(provider => new RunDiagnosticView
+                {
+                    DataContext = provider.GetRequiredService<RunDiagnosticViewModel>()
+                })
         ;
     }
 }